neofi_backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
drf-yasg
python-dotenv